% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assertions.R
\name{assert_is_project}
\alias{assert_is_project}
\title{Assert object is a sevenbridges Project}
\usage{
assert_is_project(object)
}
\arguments{
\item{object}{object that should be a sevenbridges Project object}
}
\description{
Assert object is a sevenbridges Project
}
