% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_api_connect}
\alias{cavatica_api_connect}
\title{Connect to API}
\usage{
cavatica_api_connect(from = "file", profile_name = "default")
}
\arguments{
\item{from}{where to source cavatica credentials. defualt is to look for a file (~/.sevenbridges/credentials). See README for details on how to create this file.}

\item{profile_name}{assuming credentials are coming from file, which profile should be used (text within square brackets within ~/.sevenbridges/credentials file are different profiles) (string)}
}
\value{
sevenbridges api auth object
}
\description{
Connect to the cavatica api and return the Auth object
}
