% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_api_rate_limit}
\alias{cavatica_api_rate_limit}
\title{Cavatica Api Rate Limit}
\usage{
cavatica_api_rate_limit(auth)
}
\arguments{
\item{auth}{authorisation object used to interface with cavatica api. Retrieved by running cavatica_api_connect()}
}
\value{
cavatica api rate limit (string)
}
\description{
Find the api rate limit
}
\examples{
\dontrun{
auth = cavatica_api_connect()
cavatica_api_rate_limit(auth)
}
}
