% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_app_get}
\alias{cavatica_app_get}
\title{Get Specific App}
\usage{
cavatica_app_get(project, app_id)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}
}
\value{
If on app_id is supplied, returns a sevenbridges App object. If given multiple ids, will return a list of sevenbridges App objects
}
\description{
Get sevenbridges App object using its unique ID
}
