% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_app_list_all}
\alias{cavatica_app_list_all}
\title{List all apps in project}
\usage{
cavatica_app_list_all(project)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}
}
\value{
list of all apps in project(Sevenbridges AppList Object)
}
\description{
List all apps in project
}
