% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_download_link_to_filenames}
\alias{cavatica_download_link_to_filenames}
\title{Convert download link to filename}
\usage{
cavatica_download_link_to_filenames(link)
}
\arguments{
\item{link}{cavatica file download link}
}
\value{
file names (character)
}
\description{
Convert download link to filename
}
