% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_file_from_file_id}
\alias{cavatica_file_from_file_id}
\title{Get File from ID}
\usage{
cavatica_file_from_file_id(project, file_id)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{file_id}{id of file. Can be found using api OR using the wep portal (see number after #) (string)}
}
\value{
sevenbridges File Object (or FileList if multiple IDs are supplied)
}
\description{
Get File from ID
}
