% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_file_from_filepath}
\alias{cavatica_file_from_filepath}
\title{File and FileList Operations}
\usage{
cavatica_file_from_filepath(project, filepaths, return_ids = FALSE)
}
\arguments{
\item{project}{a cavatica project object}

\item{filepaths}{the filepaths to files you want retrieve (character)}

\item{return_ids}{return file ids instead of file objects (boolean)}
}
\value{
A Files object or a list of Files Objects.
}
\description{
Finds files from filepaths: designed to be vectorised and efficient with respect to api calls
How it works is if you parse multiple filepaths with shared folders in their history,
this fuction will remembar paths to shared folders so we don't have to re-lookup the same folders over and over.
}
\details{
This is usually more efficient than spending forever building a cache which has to be updated whenever new files are added to the volume.
}
