% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_file_from_filepath2}
\alias{cavatica_file_from_filepath2}
\title{File and FileList Operations}
\usage{
cavatica_file_from_filepath2(project, filepaths)
}
\arguments{
\item{project}{a cavatica project object}

\item{filepaths}{the filepaths to files you want retrieve (character)}
}
\value{
A Files object or a list of Files Objects.
}
\description{
File and FileList Operations
}
