% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_file_id_from_files}
\alias{cavatica_file_id_from_files}
\title{Files to file IDs}
\usage{
cavatica_file_id_from_files(files)
}
\arguments{
\item{files}{sevenbridges file objects}
}
\value{
character vector describing IDs
}
\description{
Files to file IDs
}
