% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_file_path_from_file}
\alias{cavatica_file_path_from_file}
\title{cavatica_file_path_from_file}
\usage{
cavatica_file_path_from_file(project, file)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{file}{a sevenbridges file object. E.g. that prduced by cavatica_file_from_filepath}
}
\value{
path of file (string)
}
\description{
cavatica_file_path_from_file
}
