% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_project_list}
\alias{cavatica_project_list}
\title{List all projects}
\usage{
cavatica_project_list(auth = cavatica_api_connect(), project_id)
}
\arguments{
\item{auth}{authorisation object used to interface with cavatica api. Retrieved by running cavatica_api_connect()}

\item{project_id}{project id (project_owner/project_name). If unsure,  (string)}
}
\value{
sevenbridges ProjectList object
}
\description{
List all projects
}
