% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_create}
\alias{cavatica_tasks_create}
\title{Create Task}
\usage{
cavatica_tasks_create(
  project,
  name,
  app_id,
  inputs = NULL,
  batch = NULL,
  description = NULL,
  use_interruptible_instances = TRUE,
  execution_settings = NULL
)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{name}{name of task}

\item{app_id}{id of app/workflow to run}

\item{inputs}{list describing input values. See \code{\link[=cavatica_app_inputs]{cavatica_app_inputs()}} for a list of input paramaters}

\item{batch}{list describing how to batch task. Construct value passed to function using \code{\link[sevenbridges:batch]{sevenbridges::batch()}}. For example: \code{batch = batch(input = "fastq", criteria = c("metadata.sample_id","metadata.noexist_id"))}}

\item{description}{description of task}

\item{use_interruptible_instances}{use spot instances? (boolean)}

\item{execution_settings}{control the instance type and maximum number of parallel instances (list, e.g. \code{list(instance_type = "c4.2xlarge;ebs-gp2;2000", max_parallel_instances = 1, use_memoization = false, use_elastic_disk = false)})}
}
\value{
sevenbridges Task object
}
\description{
Create Task
}
