% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_delete}
\alias{cavatica_tasks_delete}
\title{Cavatica Tasks}
\usage{
cavatica_tasks_delete(tasks)
}
\arguments{
\item{tasks}{task / tasklist / list of task objects to delete}
}
\value{
run for its side effects
}
\description{
delete task / tasks.
}
