% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_download}
\alias{cavatica_tasks_download}
\title{Download cavatica task outputs}
\usage{
cavatica_tasks_download(tasks, output_dir)
}
\arguments{
\item{tasks}{task / tasklist / list of task objects to delete}

\item{output_dir}{directory to download outputfiles to. Can supply one folder for all task downloads or 1 folder for each task}
}
\value{
run for its side effects
}
\description{
Download cavatica task outputs
}
