% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_get}
\alias{cavatica_tasks_get}
\title{Get Task by ID}
\usage{
cavatica_tasks_get(project, ids)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}
}
\value{
task
}
\description{
Get Task by ID
}
