% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_list_all}
\alias{cavatica_tasks_list_all}
\title{List all tasks}
\usage{
cavatica_tasks_list_all(
  project,
  status = c("all", "queued", "draft", "running", "completed", "aborted", "failed")
)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{status}{only return tasks with the chosen status}
}
\value{
TaskList with full task list
}
\description{
List all tasks
}
