% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_run}
\alias{cavatica_tasks_run}
\title{Cavatica Tasks}
\usage{
cavatica_tasks_run(tasks)
}
\arguments{
\item{tasks}{task / tasklist / list of task objects to delete}
}
\value{
run for its side effects
}
\description{
Run cavatica tasks
}
