% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskScheduling.R
\name{cavatica_tasks_schedule}
\alias{cavatica_tasks_schedule}
\title{Task Scheduler}
\usage{
cavatica_tasks_schedule(
  project,
  task_list,
  seconds_between_job_starts,
  max_number_of_concurrent_task
)
}
\arguments{
\item{project}{cavatica project object}

\item{task_list}{a list of cavatica 'Task' objects}

\item{seconds_between_job_starts}{how long to wait after spawning a task to spawn the next one. Useful to stagger input phases.}

\item{max_number_of_concurrent_task}{Maximum number of tasks that can be running concurrently. If more than this number of tasks are running, scheduler will check every 5 minutes until enough finish before spawning any new runs}
}
\value{
NULL - run for its side effects
}
\description{
Task Scheduler
}
