% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_search}
\alias{cavatica_tasks_search}
\title{Cavatica Tasks}
\usage{
cavatica_tasks_search(
  project,
  pattern,
  status = c("all", "queued", "draft", "running", "completed", "aborted", "failed")
)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{pattern}{pattern of name e.g. 'prefix*'}

\item{status}{one of: c("all", "queued", "draft", "running", "completed", "aborted", "failed")}
}
\value{
tasklist
}
\description{
Cavatica Tasks
}
