% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_tasks_search_by_prefix}
\alias{cavatica_tasks_search_by_prefix}
\title{Cavatica Tasks}
\usage{
cavatica_tasks_search_by_prefix(
  project,
  prefix,
  status = c("all", "queued", "draft", "running", "completed", "aborted", "failed")
)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{prefix}{prefix of task name (string)}

\item{status}{one of: c("all", "queued", "draft", "running", "completed", "aborted", "failed")}
}
\value{
tasklist
}
\description{
Find tasks with name starting with 'prefix'
}
