% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{find_file_using_filepath}
\alias{find_file_using_filepath}
\title{File and FileList Operations}
\usage{
find_file_using_filepath(project, filepaths)
}
\arguments{
\item{project}{a cavatica project object}

\item{filepaths}{the filepaths to files you want retrieve (character)}
}
\value{
A Files object or a list of Files Objects.
}
\description{
File and FileList Operations
}
