% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{get_files_from_filelist_by_name}
\alias{get_files_from_filelist_by_name}
\title{File and FileList Operations}
\usage{
get_files_from_filelist_by_name(filelist, filenames)
}
\arguments{
\item{filelist}{a sevenbridges FilesList class object}

\item{filenames}{a vector of names}
}
\value{
a sevenbridges Files object OR a FileList Object (if youve supplied multiple filenames) OR NA if names don't exists in the filelist
}
\description{
File and FileList Operations
}
