% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{get_path_from_splitpath}
\alias{get_path_from_splitpath}
\title{File and FileList Operations}
\usage{
get_path_from_splitpath(project, splitpath)
}
\arguments{
\item{project}{cavatica project}

\item{splitpath}{a single vector where each element in a folder/file in a filepath}
}
\value{
File at the end of the filepath (sevenbridges Files class object) OR NA if file doesnt exist
}
\description{
File and FileList Operations
}
