% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{get_path_from_splitpath_with_explicit_parent_folder}
\alias{get_path_from_splitpath_with_explicit_parent_folder}
\title{File and FileList Operations}
\usage{
get_path_from_splitpath_with_explicit_parent_folder(
  project,
  splitpath,
  parent_folder,
  file_object_only = FALSE
)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{splitpath}{a single vector where each element in a folder/file in a filepath}

\item{parent_folder}{folder to act as 'root' folder from which splitpath is relative to (Sevenbridges File Object)}

\item{file_object_only}{return only the file object - or a list where element 1 is file object and element 2 is a dataframe of all cached filepaths stepped through on the way to finding the file (boolean)}
}
\value{
File at the end of the filepath (sevenbridges Files class object) OR NA if file doesnt exist
}
\description{
File and FileList Operations
}
