% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenStash.R
\name{standard_description_function}
\alias{standard_description_function}
\title{Roxygen Cache}
\usage{
standard_description_function(project, app_id, file_id, file, auth, status)
}
\arguments{
\item{project}{a sevenbridges 'Project' object. See readme for instructions on how to create this object}

\item{app_id}{unique app id: <project_id>/<short_app_id>/<version_number>. Can find this by running cavatica_app_list_all, cavatica_app_search, or using web platform (ID is everything after # in URL. May still need to add /<version_number>) (string)}

\item{file_id}{id of file. Can be found using api OR using the wep portal (see number after #) (string)}

\item{file}{a sevenbridges file object. E.g. that prduced by cavatica_file_from_filepath}

\item{auth}{authorisation object used to interface with cavatica api. Retrieved by running cavatica_api_connect()}

\item{status}{one of: c("all", "queued", "draft", "running", "completed", "aborted", "failed")}
}
\description{
Place to stash roxygen descriptions for common paramaters used accross package
}
