% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileQueries.R.R
\name{cavatica_file_id_from_filepath}
\alias{cavatica_file_id_from_filepath}
\title{Cavatica File From Filepath}
\usage{
cavatica_file_id_from_filepath(project, filepaths)
}
\arguments{
\item{project}{a cavatica project object}

\item{filepaths}{the filepaths to files you want retrieve (character)}
}
\value{
character filepaths
}
\description{
returns file_id from filepaths
}
