% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntacticSugar.R
\name{cavatica_project_get}
\alias{cavatica_project_get}
\title{Cavatica Get Project}
\usage{
cavatica_project_get(project_id, status = "all", auth = cavatica_api_connect())
}
\arguments{
\item{project_id}{project id (project_owner/project_name). If unsure,  (string)}

\item{status}{one of: c("all", "queued", "draft", "running", "completed", "aborted", "failed")}

\item{auth}{authorisation object used to interface with cavatica api. Retrieved by running cavatica_api_connect()}
}
\value{
cavatica project
}
\description{
Cavatica Get Project
}
